moved enzymeConfig into js/config since jest wants to run it as a test if it is in __tests__
