+++ src/__tests__/controllers/User.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../enzymeConfig';
+import 'js/config/enzymeConfig';
 import { shallow } from 'enzyme';
 
 import User from 'js/controllers/User'
